backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend App for streamlit frontend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.12",
    "openai==1.75.0",
    "pandas==2.2.3",
    "peft>=0.14.0",
    "python-dotenv==1.0.1",
    "pydantic==2.11.2",
    "torch==2.6.0",
    "transformers==5.3.0",
    "uvicorn[standard]==0.34.0",
    "pyyaml>=6.0.3",
]
